Fix memory access violations, replace boost::format
- Fix memory access violations related to the persistence path
- Replace boost::format with strprintf in mastercore.cpp
- Replace boost::format with strprintf in mastercore_sp.h
- Replace boost::format with strprintf in mastercore_sp.cp
- Minor, low hanging changes on the way

The replacement of boost::format was done, because tinyformat has similar properties, and is used in many other places of the project anyway. On top, tinyformat should be faster, but I didn't confirm this.

The combination of boost::filesystem::path, boost::format, string(), str() and c_str() caused memory related issues, according to valgrind.

Before:
- https://gist.githubusercontent.com/dexX7/d1298afb7ded4a50a129/raw/

After:
- https://gist.githubusercontent.com/dexX7/8a78b4e84a4abc3ecb24/raw/