Don't bypass transaction restrictions on non-mainnet networks
Known transaction types in the test ecosystem are still allowed.

MSC_TYPE_OFFER_ACCEPT_A_BET is activated with block 999999 on mainnet, as well as testnet and regtest mode. The unit tests show that bet transactions are not enabled, when using MSC, but allowed, when using TMSC. The tests also show that height restrictions apply on testnet, too.