Seperate logic from plain transaction parsing
To turn parseTransaction() into a side effect free function, the DEx payment and Exodus purchase logic was extracted, and moved into new functions, which are called directly from the transaction handler.

The legacy parsing is mostly untouched, and the new logic functions are only executed, if the non-legacy parsing function is used.

parseTransaction() returns a negative error code, 1 (potential DEx payment), or 0, if it's an Omni transaction with payload. This should not have an impact, given that it's usually used with checks such as < 0, == 0, >= 0, or > 0.

Because the result of the transaction handler is either valid, or invalid, a boolean is returned instead of a number.

These changes pave the way for further restructuring, and finer control over the state, as well as better signal handling.