Fix Omni transaction count value passed into block end handler
Note: Instead of passing the number of Omni transactions in the block to the block end handler, msc_initial_scan passes in nFound which is a cumulative total of the number of transactions found during the entire scan, not the number of transactions found in that particular block.  Thus the block end handler gets a positive value for the number of Omni transactions in each block regardless of whether there actually were any.  This causes any actions that should only be executed on Omni containing blocks (such as wallet cache checks) to instead be executed every block.