Reject "over-offers" after DEx feature activation
After this feature is enabled, it is no longer valid to create orders, which offer more than the seller has available, and the amounts are no longer adjusted based on the actual balance.

The second check can be left untouched, because the order adjustment is desired, as long as the feature is not enabled, and after the feature is enabled, it's body is no longer reached.