Use plain integer math for DEx purchase calculation
As long as this feature is disabled, floating point math is used to determine the purchased amount.

After this feature is enabled, plain integer math is used to determine the purchased amount. The purchased amount is rounded up, which may be in favor of the buyer, to avoid small leftovers of 1 willet.

This is not exploitable due to transaction fees.