Fix status on invalid trades
Note: prior to this patch invalid trades appear with a status of "cancelled" and include an all-zero "canceltxid" field.
This is because trades not in the live maps are assumed to be closed (and if not all sold, thus cancelled).
That assumption fails if the trade was not valid in the first place, so this patch incorporates validity when working out the status.