Ensure database objects exist, when creating iterators
When creating iterators, it is assumed that the database is not closed.

While the databases are already initialized in the constructors of the
child classes, this makes it harder to misuse the class, by restricting
who can actually close the database.

The internal database object is further initialized, as it should be,
even though this is a noop.