Add RPC to decode raw Omni transactions
The new RPC "omni_decodetransaction" can be used to decode raw Omni
transactions.

If the inputs of the transaction are not in the chain, then they must
be provided, because the transaction inputs are used to identify the
sender of a transaction.

An empty coins view is created, and populated with the user provided
transaction information. Right before the decoding, the global coins
view is temporarily swapped with the dummy view, to make it available
to "ParseTransaction". Afterwards the view is swapped back. Access to
the coins view is blocked during the decoding.