Also lock cs_main before decoding Omni transaction
The extra lock for cs_main is needed to avoid a deadlock when
connecting a block.

The lock order appears to be as follows:

1. omnidecode_transaction locks cs_tx_cache
2. Block connecting locks cs_main
3. Handlers lock cs_tally
4. Parser waits to lock cs_tx_cache
5. omnidecode_transaction can't continue, because cs_main is held

As result the RPC thread waits for cs_main, while holding cs_tx_cache,
and the core thread waits for cs_tx_cache, while holding cs_main.

Thanks to zathras-crypto for identifying the issue in the first place.