Use uint256 when calculating desired BTC for DEx 1
The legacy calculation of the desired BTC amount was not suitable for larger numbers, resulting in bad data on the RPC layer.

The legacy calculation is part of the consensus code, and used when calculating the updated desired amount, in case a seller offered more than available. This is no longer possible since 0.0.10, but the code is still needed for historical transactions.

The updated version uses plain integer math. The function is used on the RPC layer for omni_gettransaction and omni_getactivedexsells.