Use estimated fees to filter and create Omni wallet transactions
To avoid bloating transactions while lowering the effective fee, inputs are filtered during the coin selection, and uneconomic inputs are excluded. This is based on the estimated fee used for the transaction creation.

Instead of selecting an amount based on a magic number, the amount to select for the transaction creation is also based on the estimated fee, used to create the transaction.